backend/requirements.txt
python-jose
fastapi
sqlalchemy
passlib[bcrypt]
mysql-connector-python